Updated FirstWing level loading

+++ FirstWing/src/com/gamegear/firstwing/levels/json/SpeedPoint.java
@@ -1,7 +1,7 @@
 package com.gamegear.firstwing.levels.json;
 
 public class SpeedPoint {
-	public int X;
+	public float X;
 	
 	public float Speed;
 }
